Build with current compilers, and add CI
This core has not been compiled since 2018: there is no .gitlab-ci.yml, no
pipeline on the mirror and nothing on the buildbot. Most of what stops it
is age.

- -fcommon: the machine state lives in tentative definitions in the
  headers, so every translation unit owns a copy and gcc 10 stopped
  merging those.
- Demote the C89-isms gcc 14 made errors (implicit declarations, implicit
  int, int/pointer conversion) instead of rewriting the tree.
- Pick the CD-ROM backend from the target OS. Makefile.common always
  compiled cdrom-ioctl-linux.c, which includes linux/cdrom.h, so no
  Windows build could get past it; the Win32 IOCTL backend was sitting
  right next to it, unused.
- Keep the Win32 front end out of the libretro core. The #ifdef WIN32
  paths in pc.c call into nethandler.c, which we do not build, and
  video.c only stubs the D3D/DirectDraw screenshot entry points #ifdef
  __unix, so a mingw link ended in undefined references to the GUI.
- localtime_r is not linkable in mingw; use localtime there.
- Detect the 32-bit x86 recompiler: that arm filtered $(UNAME) for
  "amd64" and compared the result to "x86", which can never match, so a
  32-bit host got no recompiler and hit codegen.h's #error. Cross builds
  can now say so with $ARCH, which is how the templates ask.
- Enable SSE2 on 32-bit x86, which the Voodoo intrinsics need.
- io.h used the fixed-width types without including stdint.h.

CI covers windows-x64, windows-i686, linux-x64 and linux-i686, all four
of which build. aarch64 needs the dynarec stubbed out first - codegen.h
is an #error on every other architecture and the core files include it
unconditionally - and macOS needs a CD-ROM backend that compiles there.

+++ src/dosbox/dbopl.cc
@@ -37,6 +37,7 @@
 #include <math.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 //#include "dosbox.h"
 #include "dbopl.h"
 //#include "../save_state.h"
@@ -1437,7 +1438,7 @@ void InitTables( void ) {
 		//Add back the bits for highest ones
 		if ( i >= 16 )
 			index += 9;
-		Bitu blah = reinterpret_cast<unsigned long>( &(chip->chan[ index ]) );
+		Bitu blah = reinterpret_cast<uintptr_t>( &(chip->chan[ index ]) );
 		ChanOffsetTable[i] = blah;
 	}
 	//Same for operators
@@ -1452,7 +1453,7 @@ void InitTables( void ) {
 			chNum += 16 - 12;
 		Bitu opNum = ( i % 8 ) / 3;
 		DBOPL::Channel* chan = 0;
-		Bitu blah = reinterpret_cast<unsigned long>( &(chan->op[opNum]) );
+		Bitu blah = reinterpret_cast<uintptr_t>( &(chan->op[opNum]) );
 		OpOffsetTable[i] = ChanOffsetTable[ chNum ] + blah;
 	}
 #if 0
